- Redefined lrand48 in config_win32.h to be lbl_random since the default windows   RNG only returns 16 bit quantities.
- Modified lbl_random and lbl_srandom types to use u_int32 rather than int since
  int is only 16 bits wide on win32.

- Type conversion necessary in mbus.c from u_int32 to int32.

This is a can of worms I shouldn't have touched.

+++ src/mbus.c
@@ -108,7 +108,7 @@ static char *mbus_new_encrkey(void)
 
 	/* Step 1: generate a random string for the key... */
 	for (i = 0; i < MBUS_ENCRKEY_LEN; i++) {
-		random_string[i] = (lbl_random() | 0x000ff000) >> 24;
+		random_string[i] = ((int32)lbl_random() | 0x000ff000) >> 24;
 	}
 	/* Step 2: base64 encode that string... */
 	memset(encoded_string, 0, (MBUS_ENCRKEY_LEN*4/3)+4);
@@ -133,7 +133,7 @@ static char *mbus_new_hashkey(void)
 
 	/* Step 1: generate a random string for the key... */
 	for (i = 0; i < MBUS_HASHKEY_LEN; i++) {
-		random_string[i] = (lbl_random() | 0x000ff000) >> 24;
+		random_string[i] = ((int32)lbl_random() | 0x000ff000) >> 24;
 	}
 	/* Step 2: base64 encode that string... */
 	memset(encoded_string, 0, (MBUS_HASHKEY_LEN*4/3)+4);
